# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "mafia",
        "contact": {
            "name": "A2-LAB",
            "url": "http://a2-lab.com/"
        },
        "version": "1"
    },
    "host": "mafia.a2-lab.com",
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/api/v1/balance/": {
            "get": {
                "tags": [
                    "balance"
                ],
                "summary": "Get profile balance",
                "operationId": "balance",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, balance value provided.",
                        "schema": {
                            "type": "intefer",
                            "example": 1234.990000000000009094947017729282379150390625
                        }
                    },
                    "401": {
                        "description": "Wrong token."
                    }
                }
            }
        },
        "/api/v1/balance/details": {
            "get": {
                "tags": [
                    "balance"
                ],
                "summary": "Get details profile balance.",
                "operationId": "balance",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, balance value provided.",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/BalanceDetails"
                        }
                    },
                    "401": {
                        "description": "Wrong token."
                    }
                }
            }
        },
        "/api/v1/balance/referral": {
            "get": {
                "tags": [
                    "balance"
                ],
                "summary": "Get referal profile balance.",
                "operationId": "balance",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, balance value provided.",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/ReferralBalance"
                        }
                    },
                    "401": {
                        "description": "Wrong token."
                    }
                }
            }
        },
...
        "/api/v1/quizzes/get": {
            "get": {
                "tags": [
                    "quizzes"
                ],
                "summary": "Get quizzes",
                "operationId": "Get quizzes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, api status provided",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Quiz"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/quizzes/questions/get": {
            "get": {
                "tags": [
                    "quizzes"
                ],
                "summary": "Get quiz questions",
                "operationId": "Get quiz questions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id of quiz",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, api status provided",
                        "schema": {
                            "$ref": "#/definitions/Questions"
                        }
                    },
                    "435": {
                        "description": "This quiz isn't available now"
                    }
                }
            }
        },
        "/api/v1/quizzes/update": {
            "get": {
                "tags": [
                    "quizzes"
                ],
                "summary": "Update the quiz",
                "operationId": "Update the quiz",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "quiz_id of any question",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "score",
                        "in": "query",
                        "description": "amount of correct answers",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, api status provided"
                    },
                    "435": {
                        "description": "This quiz isn't available now"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Quiz": {
            "required": [
                "id",
                "title",
                "pivot"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "pivot": {
                    "$ref": "#/definitions/Pivot_q"
                }
            }
        },
        "Pivot_q": {
            "required": [
                "today_times",
                "limit",
                "times",
                "earned",
                "is_available"
            ],
            "properties": {
                "today_views": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "times": {
                    "type": "integer"
                },
                "earned": {
                    "type": "number"
                },
                "is_available": {
                    "type": "boolean"
                }
            }
        },
        "QuizQuestion": {
            "required": [
                "id",
                "question",
                "A",
                "B",
                "C",
                "D",
                "correct_answer"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "question": {
                    "type": "string"
                },
                "A": {
                    "type": "string"
                },
                "B": {
                    "type": "string"
                },
                "C": {
                    "type": "string"
                },
                "D": {
                    "type": "string"
                },
                "correct_answer": {
                    "type": "string"
                }
            }
        },
...
        "Questions": {
            "required": [
                "quiz_id",
                "questions"
            ],
            "properties": {
                "quiz_id": {
                    "type": "integer"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/QuizQuestion"
                    }
                }
            }
        },
...
        "BalanceDetails": {
            "properties": {
                "balance": {
                    "type": "number",
                    "format": "float",
                    "example": 1234.55999999999994543031789362430572509765625
                },
                "paid": {
                    "type": "number",
                    "format": "float",
                    "example": 1234.55999999999994543031789362430572509765625
                }
            }
        },
        "ReferralBalance": {
            "properties": {
                "referrals": {
                    "type": "object",
                    "$ref": "#/definitions/Referrals"
                },
                "balance": {
                    "type": "number",
                    "format": "float",
                    "example": 1234.55999999999994543031789362430572509765625
                },
                "award": {
                    "type": "number",
                    "format": "float",
                    "example": 20
                },
                "paid": {
                    "type": "number",
                    "format": "float",
                    "example": 1234.55999999999994543031789362430572509765625
                }
            }
        },
        "Referrals": {
            "required": [
                "current",
                "max",
                "percent"
            ],
            "properties": {
                "level": {
                    "type": "integer",
                    "example": 2
                },
                "current": {
                    "type": "integer",
                    "example": 2
                },
                "max": {
                    "type": "integer",
                    "example": 3
                },
                "percent": {
                    "type": "integer",
                    "example": 60
                }
            }
        },
...
    },
...
}